chore: fixed interrupt, to halt the system and not crash

+++ kernel/interrupts.c
@@ -6,7 +6,7 @@
 struct interrupt_descriptor_table_entry {
     uint16_t offset_low;
     uint16_t selector;
-    uint16_t zero;
+    uint8_t zero;
     uint8_t type_attr;
     uint16_t offset_high;
 }__attribute__((packed));
